Train Model Track Model Integration Working (#40)
* Semi-Working Integration

* Train Model Working with Time except temperature

* Track Model adds correct data for Trains

* Changed Paths of the csv Files

* Minor change to Integration UI Signals

* Fixing Minor issue with Service Brake

* Beacons Updates

* Track Model Train Model Integration

+++ src/TrackModel/TrackModelSignals.py
@@ -14,7 +14,7 @@ class trackModelSignals(QObject):
     getOffInput = pyqtSignal(int, int, int)
     getOnInput = pyqtSignal(int, int)
     getAuthInput = pyqtSignal(int, int)
-    getCSpeedInput = pyqtSignal(int, int)
+    getCSpeedInput = pyqtSignal(float, int)
     getRealTimeClockInput = pyqtSignal(str)
